use scalatest and some renaming

+++ src/main/scala/akka/cluster/sbr/sbr.scala
@@ -4,5 +4,5 @@ import eu.timepit.refined.api.Refined
 import eu.timepit.refined.numeric.Positive
 
 package object sbr {
-  type QuorumSize = Long Refined Positive
+  type QuorumSize = Int Refined Positive
 }
